requirements.txt
binance-futures-connector
streamlit
python-dotenv
